Migrate to frida-compile
The frida-load module is deprecated and should not be used. This is
however my fault for not making this clear in the Frida documentation,
so I'm submitting this PR to make up for it. :-)

I also replaced passionfruit_checksec() written in C with JavaScript,
by using the macho module. I did however not port the lsof functionality
as it isn't currently used by the UI. This would be easy to port though
if needed in the future.